+++ tailwind.config.js
@@ -1,0 +1,8 @@
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: ['./entrypoints/**/*.{html,ts,js}'],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+};